docker-requirements.txt
# Core Python dependencies for PF2e Society Scribe
# Python 3.10+ required

# Web Framework
starlette==0.37.2
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1

# Database
aiosqlite==0.20.0
sqlalchemy==2.0.31

# Discord Bot (Phase 5)
# discord.py==2.3.2

# Model Runtime
llama-cpp-python==0.2.90
transformers==4.44.2
tokenizers==0.19.1
huggingface-hub==0.24.5

# API and Serialization
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
orjson==3.10.6

# Utilities
invoke==2.2.0
python-dateutil==2.9.0
pytz==2024.1
aiohttp==3.10.3

# Testing (included for in-container testing)
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-mock==3.14.0

# Logging and Monitoring
structlog==24.2.0
rich==13.7.1

# MCP Server (Phase 4.5)
# Will add MCP-specific dependencies when implementing

# Development tools (optional, but useful for debugging)
ipython==8.26.0
ipdb==0.13.13